+++ data/lua/wml/random_placement.lua
@@ -14,8 +14,15 @@ wesnoth.wml_actions.random_placement = function(cfg)
 	local math_abs = math.abs
 	local locs = wesnoth.get_locations(filter)
 	if type(num_items) == "string" then
-		num_items = math.floor(load("local size = " .. #locs .. "; return " .. num_items)())
-		print("num_items=" .. num_items .. ", #locs=" .. #locs)
+		if num_items:match('^%s%(.*%)%s$') then
+			local params = {size = #locs}
+			local result = wesnoth.eval_formula(num_items)
+			num_items = math.floor(tonumber(result))
+		else
+			wesnoth.deprecated_message('num_items=lua', 2, '1.17.0', 'Use of Lua for [random_placement]num_items is deprecated. Use WFL instead and enclose the whole thing in parentheses.')
+			num_items = math.floor(load("local size = " .. #locs .. "; return " .. num_items)())
+			print("num_items=" .. num_items .. ", #locs=" .. #locs)
+		end
 	end
 	local size = #locs
 	for i = 1, num_items do